[WIN32-MSVC] Add cast operation to solve compiler warnings on x64 platform. Before this change, compiler claimed that the cast operation to wider type should be applied before other arithmetic operations because of overflow risk. There is no overflow risk here but it is modified to solve compiler warnings.

+++ FreeRTOS/Demo/WIN32-MSVC/main_full.c
@@ -463,7 +463,7 @@ void vFullDemoIdleFunction( void )
 
     /* Exercise heap_5 a bit.  The malloc failed hook will trap failed
      * allocations so there is no need to test here. */
-    pvAllocated = pvPortMalloc( ( rand() % 500 ) + 1 );
+    pvAllocated = pvPortMalloc( ( size_t )( rand() % 500 ) + 1 );
     vPortFree( pvAllocated );
 }
 /*-----------------------------------------------------------*/
